refactor: project structure
Refactors the client management code for improved organization, lazy initialization, and concurrent access, enhancing performance and preventing race conditions.

Moves client-related code into a dedicated `clients` directory, improving project structure.

Updates network details caching mechanism to use `RwLock` and `Arc` for efficient concurrent access and lazy initialization.

Improves test coverage and error handling.
