+++ crates/algokit_utils/src/clients/client_manager.rs
@@ -1,4 +1,4 @@
-use crate::network_client::{
+use crate::clients::network_client::{
     AlgoClientConfig, AlgoClients, AlgoConfig, AlgorandService, NetworkDetails, TokenHeader,
     genesis_id_is_localnet,
 };
@@ -6,7 +6,7 @@ use algod_client::AlgodClient;
 use algokit_http_client::DefaultHttpClient;
 use base64::{Engine, engine::general_purpose};
 use std::{env, sync::Arc};
-use tokio::sync::OnceCell;
+use tokio::sync::RwLock;
 
 pub struct AlgoClientsImpl {
     pub algod: AlgodClient,
@@ -22,7 +22,7 @@ impl AlgoClients for AlgoClientsImpl {
 
 pub struct ClientManager {
     clients: AlgoClientsImpl,
-    cached_network_details: OnceCell<NetworkDetails>,
+    cached_network_details: RwLock<Option<Arc<NetworkDetails>>>,
 }
 
 impl ClientManager {
@@ -33,7 +33,7 @@ impl ClientManager {
 
         Self {
             clients,
-            cached_network_details: OnceCell::new(),
+            cached_network_details: RwLock::new(None),
         }
     }
 
@@ -41,18 +41,42 @@ impl ClientManager {
         self.clients.algod()
     }
 
+    /// Most laconic, efficient, and idiomatic async lazy initialization
+    ///
+    /// Features:
+    /// - Only caches successful results (failures can be retried)
+    /// - Concurrent reads don't block each other
+    /// - Double-checked locking pattern prevents race conditions
+    /// - Returns Arc<T> for efficient shared ownership
     pub async fn network(
         &self,
-    ) -> Result<&NetworkDetails, Box<dyn std::error::Error + Send + Sync>> {
-        self.cached_network_details
-            .get_or_try_init(|| async {
-                let params = self.algod().transaction_params().await?;
-                Ok(NetworkDetails::new(
-                    params.genesis_id.clone(),
-                    general_purpose::STANDARD.encode(&params.genesis_hash),
-                ))
-            })
-            .await
+    ) -> Result<Arc<NetworkDetails>, Box<dyn std::error::Error + Send + Sync>> {
+        // Fast path: multiple readers can access concurrently
+        {
+            let cached = self.cached_network_details.read().await;
+            if let Some(ref details) = *cached {
+                return Ok(Arc::clone(details));
+            }
+        }
+
+        // Slow path: exclusive write access for initialization
+        let mut cached = self.cached_network_details.write().await;
+
+        // Double-check: someone else might have initialized while we waited for write lock
+        if let Some(ref details) = *cached {
+            return Ok(Arc::clone(details));
+        }
+
+        // Initialize - errors are NOT cached, allowing retries for transient failures
+        let params = self.algod().transaction_params().await?;
+        let network_details = Arc::new(NetworkDetails::new(
+            params.genesis_id.clone(),
+            general_purpose::STANDARD.encode(&params.genesis_hash),
+        ));
+
+        // Cache only on success
+        *cached = Some(Arc::clone(&network_details));
+        Ok(network_details)
     }
 
     pub fn genesis_id_is_localnet(genesis_id: &str) -> bool {
@@ -103,7 +127,9 @@ impl ClientManager {
     }
 
     pub fn get_algonode_config(network: &str, service: AlgorandService) -> AlgoClientConfig {
-        let subdomain = service.algonode_subdomain();
+        let subdomain = service
+            .algonode_subdomain()
+            .expect("Service must be supported on algonode");
 
         AlgoClientConfig::new(format!("https://{}-{}.algonode.cloud/", network, subdomain))
             .with_port(443)
@@ -150,5 +176,75 @@ impl ClientManager {
 
     pub fn get_algod_client_from_environment() -> AlgodClient {
         Self::get_algod_client(&Self::get_algod_config_from_environment())
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::clients::network_client::AlgorandService;
+
+    #[test]
+    fn test_cache_initially_empty() {
+        let config = AlgoConfig {
+            algod_config: AlgoClientConfig::new("http://localhost:4001".to_string()),
+        };
+        let manager = ClientManager::new(config);
+
+        // Cache should be initially empty
+        let cache = manager.cached_network_details.try_read().unwrap();
+        assert!(cache.is_none());
+    }
+
+    #[tokio::test]
+    async fn test_error_not_cached() {
+        let config = AlgoConfig {
+            algod_config: AlgoClientConfig::new("http://invalid-host:65534".to_string())
+                .with_port(65534),
+        };
+        let manager = ClientManager::new(config);
+
+        // Both calls should fail
+        assert!(manager.network().await.is_err());
+        assert!(manager.network().await.is_err());
+
+        // Cache should remain empty after errors
+        let cache = manager.cached_network_details.read().await;
+        assert!(cache.is_none());
+    }
+
+    #[test]
+    fn test_client_config_builder() {
+        let config = AlgoClientConfig::new("http://localhost".to_string())
+            .with_port(4001)
+            .with_token(TokenHeader::String("test-token".to_string()));
+
+        assert_eq!(config.server, "http://localhost");
+        assert_eq!(config.port, Some(4001));
+        assert!(matches!(config.token, Some(TokenHeader::String(_))));
+    }
+
+    #[test]
+    fn test_algonode_config() {
+        let config = ClientManager::get_algonode_config("testnet", AlgorandService::Algod);
+        assert_eq!(config.server, "https://testnet-api.algonode.cloud/");
+        assert_eq!(config.port, Some(443));
+    }
+
+    #[test]
+    fn test_localnet_config() {
+        let config = ClientManager::get_default_localnet_config(AlgorandService::Algod);
+        assert_eq!(config.server, "http://localhost");
+        assert_eq!(config.port, Some(4001));
+        assert!(config.token.is_some());
+    }
+
+    #[test]
+    fn test_genesis_id_localnet_detection() {
+        assert!(ClientManager::genesis_id_is_localnet("devnet-v1"));
+        assert!(ClientManager::genesis_id_is_localnet("sandnet-v1"));
+        assert!(ClientManager::genesis_id_is_localnet("dockernet-v1"));
+        assert!(!ClientManager::genesis_id_is_localnet("testnet-v1.0"));
+        assert!(!ClientManager::genesis_id_is_localnet("mainnet-v1.0"));
     }
 }
