+++ crates/algokit_utils/src/clients/mod.rs
@@ -1,0 +1,11 @@
+pub mod algorand_client;
+pub mod client_manager;
+pub mod network_client;
+
+// Re-export commonly used client types
+pub use algorand_client::AlgorandClient;
+pub use client_manager::ClientManager;
+pub use network_client::{
+    AlgoClientConfig, AlgoConfig, AlgorandNetwork, AlgorandService, NetworkDetails, TokenHeader,
+    genesis_id_is_localnet,
+};
