+++ crates/algokit_utils/src/transactions/mod.rs
@@ -1,0 +1,7 @@
+pub mod composer;
+
+// Re-export commonly used transaction types
+pub use composer::{
+    CommonParams, Composer, ComposerError, ComposerTxn, EmptySigner, PaymentParams, TxnSigner,
+    TxnSignerGetter,
+};
