+++ crates/algokit_utils/tests/algod/pending_transaction_information.rs
@@ -6,8 +6,12 @@ use algokit_utils::{
 };
 use std::convert::TryInto;
 
+use crate::common::init_test_logging;
+
 #[tokio::test]
 async fn test_pending_transaction_broadcast() {
+    init_test_logging();
+
     LocalnetManager::ensure_running()
         .await
         .expect("Failed to start localnet");
