+++ crates/algokit_utils/tests/clients/mod.rs
@@ -1,0 +1,1 @@
+pub mod client_manager;
