+++ crates/algokit_utils/tests/clients_tests.rs
@@ -1,0 +1,2 @@
+mod clients;
+mod common;
