--- crates/algokit_utils/tests/composer/mod.rs
@@ -1,1 +1,0 @@
-pub mod payment;
