--- crates/algokit_utils/tests/composer_tests.rs
@@ -1,2 +1,0 @@
-mod common;
-mod composer;
