+++ crates/algokit_utils/tests/transactions/mod.rs
@@ -1,0 +1,1 @@
+pub mod composer;
